Replace font icon in nav with SVG icon.
Uses `vite-plugin-svelte-svg` to replace the icon in the nav with
an SVG version.

Important: when using any of the icons in the `@material-symbols/svg-400`
package, set the viewbox to `0 0 48 48`. This is because
the SVGs are 48x48 by default. Setting the viewbox
to this allows you to scale the width and the height however
you want.