Add site settings (#29)
Adds a Sanity panel for site settings. This links to a single document
in the `siteSettings` collection with an `_id` of `siteSettings`.

Converts the site to use the settings in siteSettings where needed. Also
adds Svelte SEO, which handles setting all the meta tags for us. Adds
the ability for users to change the hero on the home page.

Closes #25.